Ensure user IDs are correctly matched in bot-related queries
Correctly cast `user_id` to `bigint` when joining `user_bots` with `users` table in `admin_bots_purchases` and `admin_bot_logs` functions within `app.py`.

+++ app.py
@@ -7852,7 +7852,7 @@ def admin_bots_purchases():
                         u.username
                     FROM user_bots ub
                     LEFT JOIN bot_types bt ON ub.bot_type = bt.bot_type
-                    LEFT JOIN users u ON ub.user_id = u.telegram_id
+                    LEFT JOIN users u ON ub.user_id::bigint = u.telegram_id
                     ORDER BY ub.created_at DESC
                     LIMIT %s OFFSET %s
                 """, (per_page, offset))
@@ -7913,7 +7913,7 @@ def admin_bot_logs(bot_id):
                         bt.price,
                         bt.icon
                     FROM user_bots ub
-                    LEFT JOIN users u ON ub.user_id = u.telegram_id
+                    LEFT JOIN users u ON ub.user_id::bigint = u.telegram_id
                     LEFT JOIN bot_types bt ON ub.bot_type = bt.bot_type
                     WHERE ub.bot_type = %s
                     ORDER BY ub.created_at DESC
